UI and sounds in place
The user interface is now in place and the basic functionality is working correctly. Flashing and sounds on section presses are also implimented.
Next need to work on the game rules and the freeplay mode. Once that is completed I would like to work on the CSS to improve the look of the interface.